requirements.txt
Django==2.0.2
psycopg2==2.7.5
pytz==2018.5
djangorestframework==3.7.7
django-filter==1.1.0
django-model-utils==3.1.1
